refactor(agent): split runtime lifecycle owners
