+++ docs/architecture/deepchat-runtime-lifecycle-owners/tasks.md
@@ -1,0 +1,9 @@
+# DeepChat Runtime Lifecycle Owners — Tasks
+
+- [x] Record the baseline and three-owner boundary.
+- [x] Extract `DeepChatLoopRunner.run()` and context-pressure recovery.
+- [x] Extract `InteractionCoordinator.respond()`.
+- [x] Extract `TurnCoordinator.start()` and `TurnCoordinator.resume()`.
+- [x] Reduce the presenter to façade, wiring, and shared leaf operations.
+- [x] Lower the architecture guard and update historical architecture documents.
+- [x] Run focused and repository validation.
